Move status to server
